#2e94 dev winnovation connect the blocks

+++ core/src/com/marama/view/controllers/DeleteObjectInputController.java
@@ -7,19 +7,45 @@ import com.marama.view.screens.GameScreen;
 
 public class DeleteObjectInputController extends InputAdapter {
     private final GameScreen gameScreen;
-
+    private SelectableInstance deletionSelected = null;
     public DeleteObjectInputController(GameScreen gameScreen) {
         this.gameScreen = gameScreen;
     }
 
+    private final float deletionOpacity = 0.5f;
+
+    @Override
+    public boolean touchDown(int screenX, int screenY, int pointer, int button) {
+        Ray ray = gameScreen.world.getPerspectiveCamera().getPickRay(screenX, screenY);
+        SelectableInstance instance = (SelectableInstance) gameScreen.world.getClosestModelInstance(ray);
+        if (instance != null) {
+            if (deletionSelected != null) {
+                deletionSelected.setOpacity(1f);
+            }
+            if (instance == deletionSelected) {
+                gameScreen.world.deleteObject(instance);
+                deletionSelected = null;
+            } else {
+                deletionSelected = ((SelectableInstance) gameScreen.world.getClosestModelInstance(ray));
+                deletionSelected.setOpacity(deletionOpacity);
+            }
+        } else {
+            if (deletionSelected != null) {
+                deletionSelected.setOpacity(1f);
+                deletionSelected = null;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public boolean touchDragged(int screenX, int screenY, int pointer) {
+        return deletionSelected != null;
+    }
     @Override
     public boolean touchUp(int screenX, int screenY, int pointer, int button) {
         Ray ray = gameScreen.world.getPerspectiveCamera().getPickRay(screenX, screenY);
         SelectableInstance instance = (SelectableInstance) gameScreen.world.getClosestModelInstance(ray);
-
-        if (instance != null) {
-            gameScreen.world.deleteObject(instance);
-        }
 
         return false; // Continue to the next 'touchDragged' listener.
     }
